fix: Initialize lateinit variable to prevent crash on add note fragment back navigation

+++ app/src/main/java/com/bmprj/planner/ui/task/TasksFragment.kt
@@ -18,16 +18,13 @@ import com.bmprj.planner.base.BaseFragment
 import com.bmprj.planner.databinding.FragmentTasksBinding
 import com.bmprj.planner.model.Task
 import com.bmprj.planner.ui.MainViewModel
-import com.google.android.material.carousel.CarouselLayoutManager
+import com.bmprj.planner.utils.makeDialogForDelete
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class TasksFragment : BaseFragment<FragmentTasksBinding>(R.layout.fragment_tasks) {
-    private val taskAdapter by lazy { TaskAdapter(::onAlarmClicked,::onTaskLongClicked, ::onCompleteClicked, ::onTaskSwiped) }
+    private val taskAdapter by lazy { TaskAdapter(::onAlarmClicked,::onTaskLongClicked, ::onCompleteClicked, ::onDeleteClicked) }
     private val taskViewModel by viewModels<TaskViewModel>()
     private val sharedViewModel by activityViewModels<MainViewModel>()
     override fun initView(view: View) {
@@ -48,7 +45,6 @@ class TasksFragment : BaseFragment<FragmentTasksBinding>(R.layout.fragment_tasks
         with(binding){
             taskListRecyclerView.layoutManager = StaggeredGridLayoutManager(2,StaggeredGridLayoutManager.VERTICAL)
             taskListRecyclerView.adapter = taskAdapter
-            taskAdapter.attachSwipeToDelete(taskListRecyclerView)
         }
     }
 
@@ -78,8 +74,18 @@ class TasksFragment : BaseFragment<FragmentTasksBinding>(R.layout.fragment_tasks
     private fun onAlarmClicked(task: Task) {
         Toast.makeText(requireContext(),task.taskDate + " " +task.taskTime, Toast.LENGTH_LONG).show()
     }
-    private fun onTaskSwiped(task: Task) {
-        taskViewModel.deleteTask(task)
+    private fun onDeleteClicked(task: Task) {
+        val alertDialog = makeDialogForDelete()
+        alertDialog.setPositiveButton("Ok"){ _,_ ->
+            taskViewModel.deleteTask(task)
+            taskAdapter.list.remove(task)
+            taskAdapter.notifyDataSetChanged()
+        }
+        alertDialog.setNegativeButton("No"){_,_ ->
+
+        }
+        alertDialog.show()
+
     }
     private fun addTaskClicked() {
         val action = TasksFragmentDirections.actionTasksFragmentToAddTaskFragment()
